+++ internal/engine/pass_signature_guard_test.go
@@ -51,6 +51,7 @@ var _ = []passSignature{
 	applyPulsarEdges,
 	applyRedisPubSubEdges,
 	applyBullMQEdges,
+	applyInngestEdges,
 	applyEventBusEdges,
 
 	// Real-time / streaming